feat: ParkingLot 엔티티에 externalId 필드 추가
