--- src/main/java/com/example/parking/domain/parking/controller/ParkingLotController.java
@@ -1,4 +1,0 @@
-package com.example.parking.domain.parking.controller;
-
-public class ParkingLotController {
-}
