--- src/main/java/com/example/parking/domain/parking/dto/ParkingLotResDto.java
@@ -1,4 +1,0 @@
-package com.example.parking.domain.parking.dto;
-
-public class ParkingLotResDto {
-}
