--- src/main/java/com/example/parking/domain/parking/entity/SpotStatus.java
@@ -1,5 +1,0 @@
-package com.example.parking.domain.parking.entity;
-
-public enum SpotStatus {
-    AVAILABLE, OCCUPIED
-}
