--- src/main/java/com/example/parking/domain/parking/entity/SpotType.java
@@ -1,5 +1,0 @@
-package com.example.parking.domain.parking.entity;
-
-public enum SpotType {
-    SMALL, LARGE, ELECTRIC
-}
