--- src/main/java/com/example/parking/domain/parking/service/ParkingLotService.java
@@ -1,4 +1,0 @@
-package com.example.parking.domain.parking.service;
-
-public class ParkingLotService {
-}
