+++ src/main/java/com/example/parking/domain/parkingLot/controller/ParkingLotController.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingLot.controller;
+
+public class ParkingLotController {
+}
