+++ src/main/java/com/example/parking/domain/parkingLot/dto/ParkingLotResDto.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingLot.dto;
+
+public class ParkingLotResDto {
+}
