+++ src/main/java/com/example/parking/domain/parkingLot/entity/SpotStatus.java
@@ -1,0 +1,5 @@
+package com.example.parking.domain.parkingLot.entity;
+
+public enum SpotStatus {
+    AVAILABLE, OCCUPIED
+}
