+++ src/main/java/com/example/parking/domain/parkingLot/entity/SpotType.java
@@ -1,0 +1,5 @@
+package com.example.parking.domain.parkingLot.entity;
+
+public enum SpotType {
+    SMALL, LARGE, ELECTRIC
+}
