+++ src/main/java/com/example/parking/domain/parkingLot/repository/ParkingLotRepository.java
@@ -1,8 +1,7 @@
-package com.example.parking.domain.parking.repository;
+package com.example.parking.domain.parkingLot.repository;
 
-import com.example.parking.domain.parking.entity.ParkingLot;
+import com.example.parking.domain.parkingLot.entity.ParkingLot;
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
 
 public interface ParkingLotRepository extends JpaRepository<ParkingLot, Long> {
     // 주차장 이름으로 검색
