+++ src/main/java/com/example/parking/domain/parkingLot/service/ParkingLotService.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingLot.service;
+
+public class ParkingLotService {
+}
