+++ src/main/java/com/example/parking/domain/parkingspot/entity/ParkingSpot.java
@@ -1,8 +1,8 @@
 package com.example.parking.domain.parkingspot.entity;
 
-import com.example.parking.domain.parking.entity.ParkingLot;
-import com.example.parking.domain.parking.entity.SpotStatus;
-import com.example.parking.domain.parking.entity.SpotType;
+import com.example.parking.domain.parkingLot.entity.ParkingLot;
+import com.example.parking.domain.parkingLot.entity.SpotStatus;
+import com.example.parking.domain.parkingLot.entity.SpotType;
 import jakarta.persistence.*;
 import lombok.*;
